example-microservices/wallet/requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.2
Django==2.1.4
django-redis==4.10.0
djangorestframework==3.9.0
idna==2.7
pycparser==2.19
PyJWT==1.7.0
PyNaCl==1.3.0
python-redis-lock==3.2.0
pytz==2018.7
redis==2.10.6
requests==2.20.1
six==1.11.0
urllib3==1.24.1
